pyproject.toml
[tool.poetry]
name = "dalle-inpainting-classification"
version = "0.1.0"
description = ""
authors = ["Fabian Jetzinger <S2110595008@fhooe.at>"]
readme = "README.md"
packages = [{include = "dalle_inpainting_classification"}]

[tool.poetry.dependencies]
python = "^3.10"
scikit-learn = "^1.2.1"
Flask = "^2.2.2"
Pillow = "^9.4.0"
requests = "^2.28.2"
torchvision = "^0.14.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"
pytest-cov = "^4.0.0"
pytest-xdist = "^3.1.0"
poethepoet = "0.18.0"

[tool.poe.tasks]
force-torch-cpu = "python -m pip install torch==1.13.1+cpu -f https://download.pytorch.org/whl/torch_stable.html"
force-torch-gpu = "python -m pip install torch==1.13.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
